adding assets|refactoring component tree|styles
