+++ frontend/src/App.jsx
@@ -1,5 +1,11 @@
-import {GetBoard} from "./components/GetBoard"
+import { GameBoard } from "./components/game/Drag"
+import { GetBoard } from "./components/game/GetBoard"
 
-const App = () =>  <GetBoard/> 
+const App = () =>  (
+  <>
+    <GetBoard/> 
+    <GameBoard/>
+  </>
+)
 
 export default App
