+++ frontend/src/components/game/GetBoard.jsx
