+++ frontend/src/components/ui/Footer.jsx
@@ -1,0 +1,1 @@
+const Footer = () => {}
