+++ frontend/src/components/ui/Header.jsx
@@ -1,0 +1,1 @@
+const Header = () => {}
