+++ frontend/src/components/ui/Layout.jsx
@@ -1,0 +1,1 @@
+export const Layout = () => {}
