+++ frontend/src/styles/drag.css
@@ -1,0 +1,29 @@
+.board-container {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	height: 100vh;
+	align-items: center;
+}
+
+#gameboard {
+	display: flex;
+}
+
+.square {
+	width: 200px;
+	height: 200px;
+}
+
+.beige {
+	background-color: #ebdbb2;
+}
+
+.brown {
+	background-color: #282828;
+}
+
+#info {
+	height: 30px;
+	justify-content: center;
+}
